+++ tests/roundtrip.rs
@@ -1,4 +1,3 @@
-use serde_json::json;
 // Deserialize each ISF, serialize it back to JSON, then deserialize it again and make sure both
 // deserialized instances match.
 #[test]
@@ -23,32 +22,4 @@ fn roundtrip_test_files() {
             assert_eq!(isf, isf2);
         }
     }
-}
-
-const OMMISSION_TEST: &str = r#"""
-/* {
-  "ISFVSN" : "2",
-  "INPUTS" : [
-    {
-      "NAME" : "inputImage",
-      "TYPE" : "image"
-    }
-  ]
- }
-*/
-void main() {
-
-}
-"""#;
-
-#[test]
-fn does_not_emit_extra_fields() {
-    let isf = isf::parse(&OMMISSION_TEST).unwrap();
-    let isf_string = serde_json::to_string_pretty(&isf).unwrap();
-    let object: serde_json::Value = serde_json::from_str(&isf_string).unwrap();
-    let expected_object = json!({
-        "ISFVSN" : "2",
-        "INPUTS" : [{ "NAME" : "inputImage", "TYPE" : "image" }]
-    });
-    assert_eq!(object, expected_object);
 }
